fix: FAQListCard 애니메이션 추가
